Detector response and projection in a virtual pad plane for glad-tpc
This initial version includes a basic detector description, detector response and a projector
on a virtual pad plane. The code is based on a preliminary version by Yelei and Anna. Macros for
the point creation and evaluation are available at macros/sim and for the projection and the
evaluation of the results in macros/proj.