Add trivial data levels for HYDRA
Mapped corresponds to the entry level filled by the Readers. It consist of a pad
identifier and an array (variable std::vector) of timebuckets with the adc info,
plus some additional control or status flags. A single mapped object per pad
with signal above threshold is expected per event, filling TClonesArray of
mapped objets per event.
Cal corresponds to the same information with calibrated adc info.
Hit corresponds to a 3D voxel in the chamber with some energy and width info, in
a way that a single cal object could create several 3D hits if multiple peaks are
detected in the adc array.